+++ packages/cursorless-engine/src/processTargets/modifiers/scopeHandlers/ScopeHandlerFactoryImpl.ts
@@ -93,7 +93,11 @@ export class ScopeHandlerFactoryImpl implements ScopeHandlerFactory {
       case "glyph":
         return new GlyphScopeHandler(this, scopeType, languageId);
       case "collectionItem":
-        return new CollectionItemScopeHandler(this, languageId);
+        return new CollectionItemScopeHandler(
+          this,
+          this.languageDefinitions,
+          languageId,
+        );
       case "surroundingPair":
         return new SurroundingPairScopeHandler(
           this.languageDefinitions,
